Move CodeMap markdown to bal-md module
